2020 Day 14 and int->long refactor.

+++ advent/Days/2015/Day07.cs
@@ -117,7 +117,7 @@ namespace advent.Days._2015
             return dict;
         }
 
-        public override int Part1(string input)
+        public override long Part1(string input)
         {
             ValueDict.Clear();
