+++ advent/Days/2015/Day14.cs
@@ -91,7 +91,7 @@ namespace advent.Days._2015
             return winningFunc(list);
         }
 
-        public override int Part1(string input)
+        public override long Part1(string input)
         {
             return Part0(input, GetWinningDistance);
         }
