+++ advent/Days/2020/Day21.cs
@@ -4,7 +4,7 @@ namespace advent.Days._2020
 {
     public class Day21 : Day
     {
-        public override int Part1(string input) => throw new NotImplementedException();
+        public override long Part1(string input) => throw new NotImplementedException();
         public override long Part2(string input) => throw new NotImplementedException();
     }
 }
