+++ advent/Days/IDay.cs
@@ -2,7 +2,7 @@
 {
     public interface IDay
     {
-        int Part1(string input);
+        long Part1(string input);
         long Part2(string input);
     }
 }
